Change options for Dl4jMlpFilter default layer
The old "has the model changed, if yes, clear layers and set the new ones" mode of setting layers didn't work: in the GUI, it's constantly destroying and creating the filter object, so the zoo model is *always* changing.

Now this just puts a flag to the user - set to true, and use the default layers, otherwise, use whatever you enter.